Dynamic navbar, project reorganizarion
