--- src/components/Home/Navbar/index.js
@@ -1,1 +1,0 @@
-export { default as Navbar} from './Navbar.jsx';
