+++ src/miscellanous/Footer/index.js
@@ -1,1 +1,1 @@
-export { default as Footer} from './Footer.jsx';
+export { default} from './Footer.jsx';
