+++ src/miscellanous/index.js
@@ -1,0 +1,2 @@
+export { default as Footer } from './Footer';
+export { default as Navbar } from './Navbar';
